Bump deps to latest
Apparently, Brave doesn't play well with WebGL because it can be used
for fingerprinting. If running fingerprinting protection in strict mode,
it breaks websites (this one included and for good reason). There's not
really a good (or any?) workaround at the moment so maybe add an alert
for Brave in the future or something?